LayoutTests/discovery/get-network-services-err-bad-type-expected.txt

Tests navigator.getNetworkServices() works in bad type case as expected.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS navigator.getNetworkServices() called err callback with bad type. Code: 2
PASS err.code is 2
PASS successfullyParsed is true

TEST COMPLETE
